+++ dvc/fs/local.py
@@ -5,8 +5,8 @@ import threading
 from fsspec import AbstractFileSystem
 from funcy import cached_property, wrap_prop
 
+from dvc.fs.system import System
 from dvc.scheme import Schemes
-from dvc.system import System
 from dvc.utils import tmp_fname
 from dvc.utils.fs import copy_fobj_to_file, copyfile, makedirs, move, remove
