+++ tests/func/test_import.py
@@ -10,8 +10,8 @@ from dvc.config import NoRemoteError
 from dvc.data.db import ODBManager
 from dvc.dvcfile import Dvcfile
 from dvc.exceptions import DownloadError, PathMissingError
+from dvc.fs.system import System
 from dvc.stage.exceptions import StagePathNotFoundError
-from dvc.system import System
 from dvc.utils.fs import makedirs, remove
 from tests.unit.fs.test_dvc import make_subrepo
 from tests.utils import clean_staging
