+++ tests/func/test_remove.py
@@ -4,8 +4,8 @@ import pytest
 
 from dvc.cli import main
 from dvc.exceptions import DvcException
+from dvc.fs.system import System
 from dvc.stage.exceptions import StageFileDoesNotExistError
-from dvc.system import System
 from dvc.utils.fs import remove
 from tests.utils import get_gitignore_content
